// apps/web/src/shared/api/gen/model/createOrder200DataCustomer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Checkout API
 * POST /api/sessions с телом {} создаёт гостевую сессию. Вставьте data.token в Authorize без слова Bearer. Суммы указаны в копейках. Данные сохраняются после перезапуска. Оплата тестовая.
 * OpenAPI spec version: 1.0.0
 */

export type CreateOrder200DataCustomer = {
  /**
   * @minLength 2
   * @maxLength 100
   * @pattern \S
   */
  name: string;
  /** @maxLength 150 */
  email: string;
  /** @pattern ^\+[1-9]\d{9,14}$ */
  phone: string;
};